added integration of miniCPM model
